fix: cast np.float64 to float in insert_signal to prevent psycopg2 schema error

+++ strat_session_orb_v1.py
@@ -193,7 +193,7 @@ def insert_signal(symbol, direction, sl_pips, tp_pips, confidence, reason):
             VALUES (%s,%s,%s,'15min',%s,%s,%s,'PENDING')
             RETURNING signal_uuid;
         """, (symbol, STRATEGY_NAME, direction,
-              round(confidence / 100, 2), round(sl_pips, 1), round(tp_pips, 1)))
+              float(round(confidence / 100, 2)), float(round(sl_pips, 1)), float(round(tp_pips, 1))))
         uuid = cur.fetchone()[0]
         conn.commit(); cur.close(); conn.close()
         msg = (f"{direction} {symbol} | SL:{sl_pips:.1f}p TP:{tp_pips:.1f}p "
